rename kubos.py to kubos-sdk.py for running in kubos-sdk container. Add kubos host script for managing kubos-sdk container
